+++ cmd/dfuseeos/cli/accounthist.go
@@ -3,7 +3,10 @@ package cli
 import (
 	"fmt"
 
+	"github.com/dfuse-io/dfuse-eosio/accounthist"
+
 	accounthistApp "github.com/dfuse-io/dfuse-eosio/accounthist/app/accounthist"
+
 	"github.com/dfuse-io/dlauncher/launcher"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
@@ -20,8 +23,9 @@ func init() {
 		RegisterFlags: func(cmd *cobra.Command) error {
 			cmd.Flags().String("accounthist-grpc-listen-addr", AccountHistGRPCServingAddr, "Address to listen for incoming gRPC requests")
 			cmd.Flags().String("accounthist-dsn", AccountHistDSN, "kvdb connection string to the accoun thistory database.")
+			cmd.Flags().String("accounthist-mode", "account", "Accounthist mode configuration. One of: 'account' or 'account-contract'")
 			cmd.Flags().Int("accounthist-shard-num", 0, "[BATCH] Shard number, between 0 and 255 inclusive. Keep default for live process")
-			cmd.Flags().Int("accounthist-max-entries-per-account", 1000, "Number of actions to keep in history for each account")
+			cmd.Flags().Int("accounthist-max-entries-per-key", 1000, "Number of actions to keep in history for each key")
 			cmd.Flags().Int("accounthist-flush-blocks-interval", 1000, "Flush to storage each X blocks.  Use 1 when live. Use a high number in batch, serves as checkpointing between restarts.")
 			cmd.Flags().Bool("accounthist-enable-injector-mode", true, "Enable mode where blocks are ingested, processed and saved to the database, when false, no write operations happen.")
 			cmd.Flags().Bool("accounthist-enable-server-mode", true, "Enable mode where the gRPC server is started and answers request(s), when false, the server is disabled and no requet(s) will be handled.")
@@ -44,17 +48,18 @@ func init() {
 			}
 
 			return accounthistApp.New(&accounthistApp.Config{
-				GRPCListenAddr:       viper.GetString("accounthist-grpc-listen-addr"),
-				KvdbDSN:              mustReplaceDataDir(dfuseDataDir, viper.GetString("accounthist-dsn")),
-				BlocksStoreURL:       mustReplaceDataDir(dfuseDataDir, viper.GetString("common-blocks-store-url")),
-				BlockstreamAddr:      blockstreamAddr,
-				EnableInjector:       viper.GetBool("accounthist-enable-injector-mode"),
-				EnableServer:         viper.GetBool("accounthist-enable-server-mode"),
-				ShardNum:             byte(shardNum),
-				MaxEntriesPerAccount: viper.GetUint64("accounthist-max-entries-per-account"),
-				FlushBlocksInterval:  flushBlocksInterval,
-				StartBlockNum:        viper.GetUint64("accounthist-start-block-num"),
-				StopBlockNum:         viper.GetUint64("accounthist-stop-block-num"),
+				KvdbDSN:             mustReplaceDataDir(dfuseDataDir, viper.GetString("accounthist-dsn")),
+				GRPCListenAddr:      viper.GetString("accounthist-grpc-listen-addr"),
+				BlocksStoreURL:      mustReplaceDataDir(dfuseDataDir, viper.GetString("common-blocks-store-url")),
+				BlockstreamAddr:     blockstreamAddr,
+				ShardNum:            byte(shardNum),
+				MaxEntriesPerKey:    viper.GetUint64("accounthist-max-entries-per-key"),
+				FlushBlocksInterval: flushBlocksInterval,
+				EnableInjector:      viper.GetBool("accounthist-enable-injector-mode"),
+				EnableServer:        viper.GetBool("accounthist-enable-server-mode"),
+				StartBlockNum:       viper.GetUint64("accounthist-start-block-num"),
+				StopBlockNum:        viper.GetUint64("accounthist-stop-block-num"),
+				AccounthistMode:     accounthist.AccounthistMode(viper.GetString("accounthist-mode")),
 			}, &accounthistApp.Modules{
 				BlockFilter: runtime.BlockFilter.TransformInPlace,
 				Tracker:     runtime.Tracker,