+++ tools/accounthist.go
@@ -3,58 +3,91 @@ package tools
 import (
 	"fmt"
 	"strconv"
+	"strings"
 
-	"go.uber.org/zap"
+	"github.com/dfuse-io/dfuse-eosio/accounthist/keyer"
 
 	"github.com/dfuse-io/dfuse-eosio/accounthist"
+
+	"github.com/dfuse-io/dfuse-eosio/accounthist/injector"
+
+	"github.com/manifoldco/promptui"
+	"go.uber.org/zap"
+
 	"github.com/dfuse-io/kvdb/store"
 	"github.com/eoscanada/eos-go"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
 
-var accounthistCmd = &cobra.Command{Use: "accounthist", Short: "Read from accout history", RunE: dmeshE}
-var accountReadCmd = &cobra.Command{Use: "read", Short: "Read an account", RunE: accountReadE, Args: cobra.ExactArgs(1)}
-var accountScanCmd = &cobra.Command{Use: "scan", Short: "Scan accounts", RunE: accountScanE}
-var accountCheckpointCmd = &cobra.Command{Use: "checkpoint", Short: "Get a shard's checkpoint", RunE: accountsCheckpointE}
+var accounthistCmd = &cobra.Command{Use: "accounthist", Short: "Read from account history"}
+var accountCmd = &cobra.Command{Use: "account", Short: "Account interactions"}
+var checkpointCmd = &cobra.Command{Use: "checkpoint", Short: "Shard checkpoint interactions", Args: cobra.ExactArgs(1), RunE: readCheckpointE}
+
+// dfuseeos tools accounthist account read {account} --dsn
+var readAccountCmd = &cobra.Command{
+	Use:   "read {account}",
+	Short: "Read an account",
+	Args:  cobra.ExactArgs(1),
+	RunE:  readAccountE,
+}
+
+// dfuseeos tools accounthist account scan --dsn
+var scanAccountCmd = &cobra.Command{
+	Use:   "scan",
+	Short: "Scan accounts",
+	RunE:  scanAccountE,
+}
+
+// dfuseeos tools accounthist account scan --dsn
+var readCheckpointCmd = &cobra.Command{
+	Use:   "read",
+	Short: "Read a shard's checkpoint",
+	Args:  cobra.ExactArgs(1),
+	RunE:  readCheckpointE,
+}
+
+var deleteCheckpointCmd = &cobra.Command{
+	Use:   "delete",
+	Short: "Delete a shard's checkpoint",
+	Args:  cobra.ExactArgs(1),
+	RunE:  deleteCheckpointE,
+}
 
 func init() {
 	Cmd.AddCommand(accounthistCmd)
-	accounthistCmd.AddCommand(accountReadCmd)
-	accounthistCmd.AddCommand(accountScanCmd)
-	accounthistCmd.AddCommand(accountCheckpointCmd)
 
-	accounthistCmd.PersistentFlags().String("dsn", "badger:///dfuse-data/kvdb/kvdb_badger.db", "KVStore DSN")
-	accountScanCmd.Flags().Int("limit", 100, "limit the number of accounts when doing scan")
+	accounthistCmd.AddCommand(accountCmd)
+	accountCmd.AddCommand(readAccountCmd, scanAccountCmd)
+
+	accounthistCmd.AddCommand(checkpointCmd)
+	checkpointCmd.AddCommand(readCheckpointCmd, deleteCheckpointCmd)
+
+	accounthistCmd.PersistentFlags().String("mode", "account", "accountgist mode one of 'account' or 'account-contract'")
+	accounthistCmd.PersistentFlags().String("dsn", "badger:///dfuse-data/kvdb/kvdb_badger.db", "kvStore DSN")
+	scanAccountCmd.Flags().Int("limit", 100, "limit the number of accounts when doing scan")
 }
 
-func accountsCheckpointE(cmd *cobra.Command, args []string) (err error) {
-
-	kvdb, err := store.New(viper.GetString("dsn"))
+func readCheckpointE(cmd *cobra.Command, args []string) (err error) {
+	kvdb, mode, err := getKVDBAndMode()
 	if err != nil {
-		return fmt.Errorf("failed to setup db: %w", err)
+		return err
 	}
-	kvdb = accounthist.NewRWCache(kvdb)
 
 	shardStr := args[0]
 	shard, err := strconv.ParseUint(shardStr, 0, 64)
 	if err != nil {
-		return fmt.Errorf("Unable to determine shard value from: %s: %w", shardStr, err)
+		return fmt.Errorf("unable to determine shard value from: %s: %w", shardStr, err)
 	}
 
-	service := accounthist.NewService(
-		kvdb,
-		nil,
-		nil,
-		byte(shard),
-		1000,
-		1,
-		0,
-		0,
-		nil,
-	)
+	switch mode {
+	case accounthist.AccounthistModeAccount:
+		injector.CheckpointKeyGenerator = keyer.EncodeAccountCheckpointKey
+	case accounthist.AccounthistModeAccountContract:
+		injector.CheckpointKeyGenerator = keyer.EncodeAccountContractCheckpointKey
+	}
 
-	checkpoint, err := service.GetShardCheckpoint(cmd.Context())
+	checkpoint, err := newService(kvdb, shard).GetShardCheckpoint(cmd.Context())
 	if err != nil {
 		return err
 	}
@@ -72,23 +105,70 @@ func accountsCheckpointE(cmd *cobra.Command, args []string) (err error) {
 	return nil
 }
 
-func accountReadE(cmd *cobra.Command, args []string) (err error) {
-	kvdb, err := store.New(viper.GetString("dsn"))
+func deleteCheckpointE(cmd *cobra.Command, args []string) (err error) {
+	kvdb, mode, err := getKVDBAndMode()
 	if err != nil {
-		return fmt.Errorf("failed to setup db: %w", err)
+		return err
 	}
-	kvdb = accounthist.NewRWCache(kvdb)
-	service := accounthist.NewService(
-		kvdb,
-		nil,
-		nil,
-		0,
-		1000,
-		1,
-		0,
-		0,
-		nil,
-	)
+
+	shardStr := args[0]
+	shard, err := strconv.ParseUint(shardStr, 0, 64)
+	if err != nil {
+		return fmt.Errorf("unable to determine shard value from: %s: %w", shardStr, err)
+	}
+
+	switch mode {
+	case accounthist.AccounthistModeAccount:
+		injector.CheckpointKeyGenerator = keyer.EncodeAccountCheckpointKey
+	case accounthist.AccounthistModeAccountContract:
+		injector.CheckpointKeyGenerator = keyer.EncodeAccountContractCheckpointKey
+	}
+
+	service := newService(kvdb, shard)
+	checkpoint, err := service.GetShardCheckpoint(cmd.Context())
+	if err != nil {
+		return err
+	}
+	if checkpoint == nil {
+		fmt.Printf("No checkpoint found for shard-%d\n", shard)
+		return nil
+	}
+	fmt.Printf("Found checkpoint for for shard-%d\n", shard)
+	fmt.Printf("  - Initial Start Block for shard: %d\n", checkpoint.InitialStartBlock)
+	fmt.Printf("  - Target Stop Block for shard: %d\n", checkpoint.TargetStopBlock)
+	fmt.Printf("  - Last Written Block Num for shard: %d\n", checkpoint.LastWrittenBlockNum)
+	fmt.Printf("  - Last Written Block Id for shard: %s\n", checkpoint.LastWrittenBlockId)
+
+	prompt := promptui.Prompt{
+		Label:     "Are you sure you want to delete the checkpoint?",
+		IsConfirm: true,
+	}
+
+	result, err := prompt.Run()
+	if err != nil && err != promptui.ErrAbort {
+		fmt.Printf("Error: \n")
+	}
+
+	if strings.ToLower(result) != "y" {
+		fmt.Printf("aborting deletion. Goodbye!\n")
+		return nil
+	}
+
+	err = service.DeleteCheckpoint(cmd.Context(), byte(shard))
+	if err != nil {
+		return fmt.Errorf("unable to delete checkpoint: %w", err)
+	}
+
+	fmt.Printf("checkpoint deleted successfully!\n")
+
+	return nil
+}
+
+func readAccountE(cmd *cobra.Command, args []string) (err error) {
+	kvdb, mode, err := getKVDBAndMode()
+	if err != nil {
+		return err
+	}
 
 	account := args[0]
 	accountUint, err := eos.StringToName(account)
@@ -96,10 +176,18 @@ func accountReadE(cmd *cobra.Command, args []string) (err error) {
 		return fmt.Errorf("unable to encode string %s to eos name (utin64): %w", account, err)
 	}
 
+	switch mode {
+	case accounthist.AccounthistModeAccount:
+		injector.InjectorRowKeyDecoder = accounthist.AccountKeyRowDecoder
+	case accounthist.AccounthistModeAccountContract:
+		injector.InjectorRowKeyDecoder = accounthist.AccountContractKeyRowDecoder
+	}
+
 	zlog.Info("retrieving shard summary for account",
 		zap.String("account", account),
 	)
-	summary, err := service.ShardSummary(cmd.Context(), accountUint)
+
+	summary, err := newService(kvdb, 0).KeySummary(cmd.Context(), accounthist.AccountKey(accountUint))
 	if err != nil {
 		return fmt.Errorf("unable to retrieve account summary: %w", err)
 	}
@@ -116,28 +204,28 @@ func accountReadE(cmd *cobra.Command, args []string) (err error) {
 
 }
 
-func accountScanE(cmd *cobra.Command, args []string) (err error) {
+func scanAccountE(cmd *cobra.Command, args []string) (err error) {
 	scanLimit := viper.GetInt("limit")
-	kvdb, err := store.New(viper.GetString("dsn"))
+	kvdb, mode, err := getKVDBAndMode()
 	if err != nil {
-		return fmt.Errorf("failed to setup db: %w", err)
+		return err
 	}
-	kvdb = accounthist.NewRWCache(kvdb)
-	service := accounthist.NewService(
-		kvdb,
-		nil,
-		nil,
-		0,
-		1000,
-		1,
-		0,
-		0,
-		nil,
-	)
+	kvdb = injector.NewRWCache(kvdb)
+
+	var prefix byte
+	var decoder accounthist.RowKeyDecoderFunc
+	switch mode {
+	case accounthist.AccounthistModeAccount:
+		prefix = keyer.PrefixAccount
+		decoder = accounthist.AccountKeyRowDecoder
+	case accounthist.AccounthistModeAccountContract:
+		prefix = keyer.PrefixAccountContract
+		decoder = accounthist.AccountContractKeyRowDecoder
+	}
 
 	fmt.Printf("Scanning accounts (limit: %d)\n", scanLimit)
 	count := 0
-	err = service.ScanAccounts(cmd.Context(), func(account uint64, shard byte, ordinalNum uint64) error {
+	err = accounthist.ScanAccounts(cmd.Context(), kvdb, prefix, decoder, func(account uint64, shard byte, ordinalNum uint64) error {
 		if count > scanLimit {
 			return fmt.Errorf("scan limit reached")
 		}
@@ -150,4 +238,36 @@ func accountScanE(cmd *cobra.Command, args []string) (err error) {
 	}
 	return nil
 
+}
+
+func newService(kvdb store.KVStore, shardNum uint64) *injector.Injector {
+	return injector.NewInjector(
+		injector.NewRWCache(kvdb),
+		nil,
+		nil,
+		byte(shardNum),
+		1000,
+		1,
+		0,
+		0,
+		nil,
+	)
+
+}
+
+func getKVDBAndMode() (store.KVStore, accounthist.AccounthistMode, error) {
+	kvdb, err := store.New(viper.GetString("dsn"))
+	if err != nil {
+		return nil, "", fmt.Errorf("failed to setup db: %w", err)
+	}
+
+	switch accounthist.AccounthistMode(viper.GetString("mode")) {
+	case accounthist.AccounthistModeAccount:
+		return kvdb, accounthist.AccounthistModeAccount, nil
+	case accounthist.AccounthistModeAccountContract:
+		return kvdb, accounthist.AccounthistModeAccountContract, nil
+	default:
+		return nil, "", fmt.Errorf("unknown acounthist mode: %s", viper.GetString("mode"))
+
+	}
 }